Pramida|#759|REST API call to delete image or video file should be called only if the obs type is video or image and the obs have a value

+++ ui/app/common/domain/services/visitDocumentService.js
@@ -4,7 +4,7 @@ angular.module('bahmni.common.domain')
     .service('visitDocumentService', ['$http', 'auditLogService', 'configurations', '$q', function ($http, auditLogService, configurations, $q) {
         var removeVoidedDocuments = function (documents) {
             documents.forEach(function (document) {
-                if (document.voided) {
+                if (document.voided && document.image) {
                     var url = Bahmni.Common.Constants.RESTWS_V1 + "/bahmnicore/visitDocument?filename=" + document.image;
                     $http.delete(url, {withCredentials: true});
                 }
